Move array utils form KiwiReflection2 to KiwiArrays2
These utilities should not have gone into KiwiReflection2.
Instead, they should go in a utility class specific for arrays.
Created KiwiArrays2 and moved emptyArray and newArray there.
